Fix for Authentication Token Never Expiring
Originally using the DateTime.Now.Milliseconds computing would never allow the token to expire in an extended session.  I switched the expiresIn variable to a DateTime and used a comparison between two DateTimes to determine if we need a new token.  This fixed the issue and properly regenerated the auth token after it expired.